refactor: group compat files and move more files to deprecated
