+++ packages/backend-common/src/alpha.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export * from './context';
+export * from './deprecated/context';
