+++ packages/backend-common/src/compat/auth/createLegacyAuthAdapters.test.ts
