+++ packages/backend-common/src/compat/legacy/legacy.test.ts
@@ -21,12 +21,12 @@ import {
 import { mockServices, startTestBackend } from '@backstage/backend-test-utils';
 import { EventEmitter } from 'events';
 import { Router } from 'express';
-import { createLegacyAuthAdapters } from './auth';
+import { createLegacyAuthAdapters } from '..';
 import { legacyPlugin } from './legacy';
 // eslint-disable-next-line @backstage/no-relative-monorepo-imports
-import { authServiceFactory } from '../../backend-app-api/src/services/implementations/auth';
+import { authServiceFactory } from '../../../../backend-app-api/src/services/implementations/auth';
 // eslint-disable-next-line @backstage/no-relative-monorepo-imports
-import { tokenManagerServiceFactory } from '../../backend-app-api/src/services/implementations/tokenManager';
+import { tokenManagerServiceFactory } from '../../../../backend-app-api/src/services/implementations/tokenManager';
 
 describe('legacyPlugin', () => {
   it('can auth across the new and old systems', async () => {
