+++ packages/backend-common/src/index.ts
@@ -20,13 +20,7 @@
  * @packageDocumentation
  */
 
-export { legacyPlugin, makeLegacyPlugin } from './legacy';
-export type { LegacyCreateRouter } from './legacy';
-export { loadBackendConfig } from './config';
 export * from './deprecated';
-export * from './auth';
-export * from './cache';
-export * from './hot';
-export * from './logging';
+export * from './compat';
 export * from './middleware';
 export * from './service';
